requirements.txt
selenium
Pillow
opencv-python
numpy
requests
PyQt6
orjson
lxml
svgpathtools
webdriver-manager
PyQt6-WebEngine
# Optional modern manga OCR backend (requires torch & model download):
# Install only if you want the new `manga_ocr` backend
manga-ocr